- add rest of the platforms effects
