+++ src/Controls/samples/Controls.Sample/Pages/Core/EffectsPage.xaml
@@ -2,7 +2,7 @@
 <views:BasePage
     xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
     xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
-    x:Class="Maui.Controls.Sample.Pages.Effects"
+    x:Class="Maui.Controls.Sample.Pages.EffectsPage"
     xmlns:views="clr-namespace:Maui.Controls.Sample.Pages.Base"
     xmlns:local="clr-namespace:Maui.Controls.Sample.Pages"
     Title="Alerts">
